Splitting FileSyncTfs into sync and file handler.
