+++ src/Rainbow/SourceControl/FileSyncDefault.cs
@@ -2,17 +2,17 @@ namespace Rainbow.SourceControl
 {
 	public class FileSyncDefault : ISourceControlSync
 	{
-		public bool CheckoutFileForDelete()
+		public bool CheckoutFileForDelete(string filename)
 		{
 			return true;
 		}
 
-		public bool CheckoutFileForEdit()
+		public bool CheckoutFileForEdit(string filename)
 		{
 			return true;
 		}
 
-		public bool AddFile()
+		public bool AddFile(string filename)
 		{
 			return true;
 		}
