+++ src/Rainbow/SourceControl/ISourceControlSync.cs
@@ -2,8 +2,8 @@ namespace Rainbow.SourceControl
 {
 	public interface ISourceControlSync
 	{
-		bool CheckoutFileForDelete();
-		bool CheckoutFileForEdit();
-		bool AddFile();
+		bool CheckoutFileForDelete(string filename);
+		bool CheckoutFileForEdit(string filename);
+		bool AddFile(string filename);
 	}
 }
