+++ src/Rainbow/SourceControl/ScmSettings.cs
@@ -2,9 +2,9 @@ namespace Rainbow.SourceControl
 {
 	public class ScmSettings
 	{
-		public string Filename { get; set; }
 		public string Username { get; set; }
 		public string Domain { get; set; }
 		public string Password { get; set; }
+		public string ApplicationRootPath { get; set; }
 	}
 }
